# docs/swagger.yaml
definitions:
...
  dto.EntidadeDocumentoListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/dto.EntidadeDocumentoResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.EntidadeDocumentoRequest:
    properties:
      arquivo:
        description: Base64
        type: string
      created_by:
        description: |-
          ============================================================
          AUDITORIA
          ============================================================
        type: integer
      data_inclusao:
        description: 'Formato: 2006-01-02'
        type: string
      descricao:
        type: string
      entidade_id:
        description: |-
          ============================================================
          CAMPOS PRINCIPAIS
          ============================================================
        type: integer
      tipo:
        type: string
      updated_by:
        type: integer
    required:
    - arquivo
    - data_inclusao
    - entidade_id
    type: object
  dto.EntidadeDocumentoResponse:
    properties:
      arquivo:
        description: Base64 (opcional)
        type: string
      content_type:
        description: 'Tipo do arquivo (ex: application/pdf)'
        type: string
      created_at:
        description: |-
          ============================================================
          AUDITORIA
          ============================================================
        type: string
      created_by:
        type: integer
      data_inclusao:
        type: string
      descricao:
        type: string
      entidade_id:
        description: |-
          ============================================================
          CAMPOS PRINCIPAIS
          ============================================================
        type: integer
      item:
        description: Sequencial por entidade
        type: integer
      tamanho:
        description: |-
          ============================================================
          METADADOS
          ============================================================
        type: integer
      tipo:
        type: string
      tipo_label:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /entidades/{entidadeId}/documentos:
    get:
      description: Retorna todos os documentos de uma entidade
      parameters:
      - description: ID da entidade
        in: path
        name: entidadeId
        required: true
        type: integer
      - default: 10
        description: Limite de registros
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por tipo de documento
        in: query
        name: tipo
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EntidadeDocumentoListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Lista documentos de uma entidade
      tags:
      - EntidadeDocumentos
    post:
      consumes:
      - application/json
      description: Cadastra um novo documento para uma entidade
      parameters:
      - description: ID da entidade
        in: path
        name: entidadeId
        required: true
        type: integer
      - description: Dados do documento
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EntidadeDocumentoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.EntidadeDocumentoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cria um novo documento
      tags:
      - EntidadeDocumentos
  /entidades/{entidadeId}/documentos/{item}:
    delete:
      description: Realiza a exclusão lógica de um documento
      parameters:
      - description: ID da entidade
        in: path
        name: entidadeId
        required: true
        type: integer
      - description: ID do documento
        in: path
        name: item
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Exclui um documento
      tags:
      - EntidadeDocumentos
    get:
      description: Retorna os dados de um documento específico
      parameters:
      - description: ID da entidade
        in: path
        name: entidadeId
        required: true
        type: integer
      - description: ID do documento
        in: path
        name: item
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EntidadeDocumentoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Busca um documento específico
      tags:
      - EntidadeDocumentos
    put:
      consumes:
      - application/json
      description: Atualiza os dados de um documento existente
      parameters:
      - description: ID da entidade
        in: path
        name: entidadeId
        required: true
        type: integer
      - description: ID do documento
        in: path
        name: item
        required: true
        type: integer
      - description: Dados atualizados
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.EntidadeDocumentoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.EntidadeDocumentoResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Atualiza um documento
      tags:
      - EntidadeDocumentos
  /entidades/{entidadeId}/documentos/{item}/download:
    get:
      description: Download do arquivo de um documento
      parameters:
      - description: ID da entidade
        in: path
        name: entidadeId
        required: true
        type: integer
      - description: ID do documento
        in: path
        name: item
        required: true
        type: integer
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Arquivo do documento
          schema:
            type: file
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Download do arquivo
      tags:
      - EntidadeDocumentos
...
swagger: "2.0"